requirements.txt
python-binance==1.0.19
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14
TA-Lib==0.4.28
optuna==3.6.1
pyswarms==1.3.0
xgboost==2.0.3
scikit-learn==1.4.2
shap==0.44.1
backtrader==1.9.78.123
mlflow==2.12.1
duckdb==0.10.2
pyarrow==16.1.0
polars==0.20.24
python-dotenv==1.0.1
requests==2.31.0
protobuf==5.26.1 